+++ scraper/src/helper.ts
