+++ scraper/src/interfaces.ts
