+++ scraper/src/scraper.ts
@@ -13,7 +13,7 @@ import {
   getUrlsParams,
   scrapePage,
   getTermFromCourse,
-} from './PageScraper'
+} from './PageScraper/PageScraper'
 import { keysOf, createPages } from './helper'
 
 interface ScrapeSubjectParams {
@@ -102,9 +102,12 @@ const timetableScraper = async (
   { timetableData: TimetableData; courseWarnings: CourseWarning[] } | false
 > => {
   // Launch the browser. Headless mode = true by default
-  const browser = await puppeteer.launch({ 
+  const browser = await puppeteer.launch({
     headless: process.env.NODE_ENV === 'DEV' ? false : true,
-    args: process.env.NODE_ENV === 'DEV' ? undefined : ['--no-sandbox', '--disable-setuid-sandbox', '--disable-gpu']
+    args:
+      process.env.NODE_ENV === 'DEV'
+        ? undefined
+        : ['--no-sandbox', '--disable-setuid-sandbox', '--disable-gpu'],
   })
   try {
     const batchsize = 50
